* use 4 bytes instead of 2 in EternalFileSystemFatEntry. Increase cluster size from 512B to 4096B.

+++ EternalFS/EternalFS.Library/Filesystem/Accessors/EternalFileSystemManager.cs
@@ -146,7 +146,7 @@ public class EternalFileSystemManager : AccessorPipelineElement
 
             while (fatRef != EternalFileSystemMounter.FatTerminator)
             {
-                int offset = EternalFileSystemHelper.GetFatEntryOffset(fatRef);
+                long offset = EternalFileSystemHelper.GetFatEntryOffset(fatRef);
 
                 stream.Seek(offset, SeekOrigin.Begin);
                 fatRef = stream.MarshalReadStructure<EternalFileSystemFatEntry>();
